# crates/acton-localnet-ui/public/openapi/ton-api-v3.openapi.json
{
  "schemes": [],
  "swagger": "2.0",
  "info": {
    "description": "TON Index collects data from a full node to PostgreSQL database and provides convenient API to an indexed blockchain.\n\nWhen Acton localnet is started with --require-auth, pass the printed localnet API token as `X-API-Key` or `Authorization: Bearer <token>`.\n\nThis bundled reference is derived from the official TonCenter v3 OpenAPI document and filtered to endpoints implemented by Acton localnet.",
    "title": "Acton Localnet TonCenter API v3",
    "contact": {},
    "version": "1.2.6"
  },
  "host": "",
  "basePath": "",
  "paths": {
    "/api/v3/blocks": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Get blocks by specified filter.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["blockchain"],
        "summary": "Get Blocks",
        "operationId": "api_v3_get_blocks",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "Block workchain.",
            "name": "workchain",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Block shard id. Must be sent with *workchain*. Example: `8000000000000000`.",
            "name": "shard",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int32",
            "description": "Block seqno. Must be sent with *workchain* and *shard*.",
            "name": "seqno",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Block root hash.",
            "name": "root_hash",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Block file hash.",
            "name": "file_hash",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int32",
            "description": "Masterchain block seqno. For masterchain blocks this matches block seqno; for workchain blocks this matches the referenced masterchain block.",
            "name": "mc_seqno",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "description": "Query blocks generated after the given UTC timestamp.",
            "name": "start_utime",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "description": "Query blocks generated before the given UTC timestamp.",
            "name": "end_utime",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Query blocks with start logical time greater than or equal to the given value.",
            "name": "start_lt",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Query blocks with start logical time less than or equal to the given value.",
            "name": "end_lt",
            "in": "query"
          },
          {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "default": 10,
            "description": "Limit number of queried rows. Use with *offset* to batch read.",
            "name": "limit",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 0,
            "description": "Skip first N rows. Use with *limit* to batch read.",
            "name": "offset",
            "in": "query"
          },
          {
            "enum": ["asc", "desc"],
            "type": "string",
            "default": "desc",
            "description": "Sort blocks by generation time and seqno.",
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BlocksResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
    "/api/v3/traces": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Get traces by specified filter.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["actions"],
        "summary": "Get Traces",
        "operationId": "api_v3_get_traces",
        "parameters": [
          {
            "type": "string",
            "description": "List of account addresses to get transactions. Can be sent in hex, base64 or base64url form.",
            "name": "account",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "Find trace by trace id.",
            "name": "trace_id",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "Find trace by transaction hash.",
            "name": "tx_hash",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "Find trace by message hash.",
            "name": "msg_hash",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int32",
            "description": "Query traces that was completed in masterchain block with given seqno",
            "name": "mc_seqno",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "description": "Query traces, which was finished **after** given timestamp.",
            "name": "start_utime",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "description": "Query traces, which was finished **before** given timestamp.",
            "name": "end_utime",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Query traces with `end_lt >= start_lt`.",
            "name": "start_lt",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Query traces with `end_lt <= end_lt`.",
            "name": "end_lt",
            "in": "query"
          },
          {
            "type": "boolean",
            "default": false,
            "description": "Include trace actions.",
            "name": "include_actions",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "Supported action types",
            "name": "supported_action_types",
            "in": "query"
          },
          {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "default": 10,
            "description": "Limit number of queried rows. Use with *offset* to batch read.",
            "name": "limit",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 0,
            "description": "Skip first N rows. Use with *limit* to batch read.",
            "name": "offset",
            "in": "query"
          },
          {
            "enum": ["asc", "desc"],
            "type": "string",
            "default": "desc",
            "description": "Sort traces by lt.",
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TracesResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
    "/api/v3/accountStates": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Query account states",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["accounts"],
        "summary": "Get Account States",
        "operationId": "api_v3_get_account_states",
        "parameters": [
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "List of addresses in any form. Maximum 1000 addresses allowed.",
            "name": "address",
            "in": "query",
            "required": true
          },
          {
            "type": "boolean",
            "default": true,
            "description": "Include code and data BOCs. Default: true",
            "name": "include_boc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AccountStatesResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
    "/api/v3/addressInformation": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Get smart contract information.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["api/v2"],
        "summary": "Get Address Information",
        "operationId": "api_v3_get_v2_addressInformation",
        "parameters": [
          {
            "type": "string",
            "description": "Account address in any form.",
            "name": "address",
            "in": "query",
            "required": true
          },
          {
            "type": "boolean",
            "default": true,
            "description": "Use method from api/v2. Not recommended",
            "name": "use_v2",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/V2AddressInformation"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
...
    "/api/v3/transactionsByMessage": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Get transactions whose inbound/outbound message has the specified hash. \\",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["blockchain"],
        "summary": "Get transactions by message",
        "operationId": "api_v3_get_transactions_by_message",
        "parameters": [
          {
            "type": "string",
            "description": "Message hash. Acceptable in hex, base64 and base64url forms.",
            "name": "msg_hash",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Hash of message body.",
            "name": "body_hash",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Opcode of message in hex or signed 32-bit decimal form.",
            "name": "opcode",
            "in": "query"
          },
          {
            "enum": ["in", "out"],
            "type": "string",
            "description": "Direction of message.",
            "name": "direction",
            "in": "query"
          },
          {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "default": 10,
            "description": "Limit number of queried rows. Use with *offset* to batch read.",
            "name": "limit",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 0,
            "description": "Skip first N rows. Use with *limit* to batch read.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TransactionsResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
    "/api/v3/pendingTransactions": {
      "get": {
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "description": "Get pending transactions by specified filter.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["blockchain"],
        "summary": "Get pending transactions",
        "operationId": "api_v3_get_pending_transactions",
        "parameters": [
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "List of account addresses to get transactions. Can be sent in hex, base64 or base64url form.",
            "name": "account",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "Find transactions by trace id.",
            "name": "trace_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TransactionsResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/RequestError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AccountStatesResponse": {
      "type": "object",
      "properties": {
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountStateFull"
          }
        },
        "address_book": {
          "$ref": "#/definitions/AddressBook"
        },
        "metadata": {
          "$ref": "#/definitions/Metadata"
        }
      }
    },
...
    "AddressBook": {
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/AddressBookRow"
      }
    },
...
    "BlocksResponse": {
      "type": "object",
      "properties": {
        "blocks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Block"
          }
        }
      }
    },
...
    "Metadata": {
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/AddressMetadata"
      }
    },
...
    "RequestError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "TracesResponse": {
      "type": "object",
      "properties": {
        "address_book": {
          "$ref": "#/definitions/AddressBook"
        },
        "metadata": {
          "$ref": "#/definitions/Metadata"
        },
        "traces": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Trace"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "name": "X-Api-Key",
      "in": "header"
    },
    "APIKeyQuery": {
      "type": "apiKey",
      "name": "api_key",
      "in": "query"
    }
  },
...
}